refactor: QuertLog 완전 삭제 및 기존 조회 쿼리는 BlockTickerRepository 로 책임 위임
