+++ src/main/java/com/joojoo/api/blockTicker/domain/repository/BlockTickerRepository.java
@@ -4,6 +4,7 @@ import com.joojoo.api.blockTicker.domain.model.entity.BlockTicker;
 import com.joojoo.api.blockTicker.presentation.dto.request.TickerDateResult;
 
 import java.time.LocalDate;
+import java.util.Collection;
 import java.util.List;
 
 public interface BlockTickerRepository {
@@ -15,4 +16,7 @@ public interface BlockTickerRepository {
 
     /** 2일치 날짜 조회해서 사용한 티커들 조회 */
     TickerDateResult findAllByTickerAndDate(Long userId, Long tickerId, LocalDate targetDate);
+
+    /** TradeLog에서 사용한 티커들 조회 */
+    List<BlockTicker> findAllTickersByTradeLogIds(List<Long> tradeLogIds);
 }
