+++ src/main/java/com/joojoo/api/util/detailQuery/service/BlockDataFetcher.java
@@ -1,7 +1,6 @@
 package com.joojoo.api.util.detailQuery.service;
 
 import com.joojoo.api.block.domain.repository.BlockRepository;
-import com.joojoo.api.blockTag.application.test.QueryDsl;
 import com.joojoo.api.blockTag.domain.repository.BlockTagRepository;
 import com.joojoo.api.blockTicker.domain.repository.BlockTickerRepository;
 import com.joojoo.api.util.detailQuery.dto.BlockRelatedDataBundle;
@@ -17,7 +16,6 @@ public class BlockDataFetcher {
     private final BlockTagRepository blockTagRepository;
     private final BlockTickerRepository blockTickerRepository;
     private final BlockRepository blockRepository;
-    private final QueryDsl queryDsl;
 
     public BlockRelatedDataBundle fetchRelatedData(List<Long> blockIds, List<Long> tradeLogIds) {
         return new BlockRelatedDataBundle(
@@ -32,7 +30,7 @@ public class BlockDataFetcher {
                 blockTagRepository.findAllTagsByTradeLogIds(tradeLogIds).stream()
                         .collect(Collectors.groupingBy(bt -> bt.getTradeLog().getId())),
 
-                queryDsl.findAllTickersByTradeLogIds(tradeLogIds).stream()
+                blockTickerRepository.findAllTickersByTradeLogIds(tradeLogIds).stream()
                         .collect(Collectors.groupingBy(bt -> bt.getTradeLog().getId()))
         );
     }
